+++ DanceHub/DanceHub.csproj
@@ -177,10 +177,21 @@
     <Compile Include="Migrations\201608231727136_InitialModel.Designer.cs">
       <DependentUpon>201608231727136_InitialModel.cs</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\201608231752500_CreateDanceTable.cs" />
+    <Compile Include="Migrations\201608231752500_CreateDanceTable.Designer.cs">
+      <DependentUpon>201608231752500_CreateDanceTable.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Migrations\201608231818592_OverrideConventionsForEventsAndTypes.cs" />
+    <Compile Include="Migrations\201608231818592_OverrideConventionsForEventsAndTypes.Designer.cs">
+      <DependentUpon>201608231818592_OverrideConventionsForEventsAndTypes.cs</DependentUpon>
+    </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
-    <Compile Include="Models\IdentityModels.cs" />
+    <Compile Include="Models\ApplicationDbContext.cs" />
+    <Compile Include="Models\Event.cs" />
+    <Compile Include="Models\ApplicationUser.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
+    <Compile Include="Models\Type.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
@@ -254,6 +265,12 @@
   <ItemGroup>
     <EmbeddedResource Include="Migrations\201608231727136_InitialModel.resx">
       <DependentUpon>201608231727136_InitialModel.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201608231752500_CreateDanceTable.resx">
+      <DependentUpon>201608231752500_CreateDanceTable.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201608231818592_OverrideConventionsForEventsAndTypes.resx">
+      <DependentUpon>201608231818592_OverrideConventionsForEventsAndTypes.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <PropertyGroup>
