+++ DanceHub/Models/ApplicationDbContext.cs
@@ -1,0 +1,21 @@
+﻿using System.Data.Entity;
+using Microsoft.AspNet.Identity.EntityFramework;
+
+namespace DanceHub.Models
+{
+    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+    {
+        public DbSet<Event> Events { get; set; }
+        public DbSet<Type> Types { get; set; }
+
+        public ApplicationDbContext()
+            : base("DefaultConnection", throwIfV1Schema: false)
+        {
+        }
+
+        public static ApplicationDbContext Create()
+        {
+            return new ApplicationDbContext();
+        }
+    }
+}
